+++ CHANGES.md
@@ -1,6 +1,11 @@
 
 # Release Notes  
 
+* **ENHANCEMENT:** For user convenience, when you switch to a new session, the symbolic links found within the workspace root folder are automatically switched to point to the current session assets (`products`, `qc`, `sof`, `soxspipe.yaml`,`soxspipe.db` etc). 
+
+* **REFACTOR:** `clipping-lower-sigma` and `clipping-iteration-count` renamed to `frame-clipping-sigma` and `frame-clipping-iterations`
+* **REFACTOR:** optimised `qc_periodic_pattern_noise` for speed
+* **REFACTOR:** mask from the sigma-clipped noise map is added to the master bias frames
 
 ## v0.9.7 - December 7, 2023